A bunch of changes:
Add different date sorting options to gmail.

Fix an issue with the back button disappearing on the conversation view.

Add an archive action to emails and a button for it in the conversation
view.

Use the syncing dao again.

Make the back button of the conversation view consistent with the back
button in the search toolbar.

Make use of format=minimal when forward syncing.

Clear the search query the search view in the AppController.

Add a trueClass and falseClass to the ImageBooleanView.