Add a generic foam.core.dao.Sync class which will synchronize two DAOs.
DAOs are required to provide versioning of objects so it can easily determine which objects are out of date.  This can generally be added easily with the VersionNoDAO decorator.

A properly configured MergeDAO decorator is recommended to handle conflicts.

The protocol is "stateless" in that all necessary data is already stored in the DAOs, so sync can be resumed and continued at any time so long as the DAOs are accessible.