Java: Add ArrayProperty, ReferenceProperty, and ReferenceArrayProperty support.
Arrays are stored as Lists in Java. Native arrays are awkward and lame, and
there's no "anything[]". Object[] is not a superclass of int[], and they
otherwise have no parent class. The really fatal problem is that they can't be
resized and therefore cannot be treated as (writable) DAOs.

Also refactored JavaSource to split out property and relationship codegen,
making property codegen much, much more consistent and clear.