WebsocketDAO and corresponding node.js changes.
WebsocketDAO is an EasyClientDAO that also opens a websocket
channel to the server to relay put, remove, and reset events
from the server side DAO to the client side. The effet of this
is that if any user modifies the DAO all users are notified
so that all the client UIs update live.