Auth: Rename Authenticator -> Authorizer and AuthenticatedDAO -> AuthorizingDAO
That includes the tests and the PrivateOwner and PublicReadPrivateWrite stock
implementations of the Authorizer interface.

This is the correct meaning. GoogleAuthDAO and similar is for *authentication*:
determining a user is who they claim to be. AuthorizingDAO and its Authorizer
delegate are responsible for the other half of the "auth" problem: determining
what this user is *authorized* to do.