U2: Change the CSS shorthand from $ to ^
I think I got all the places where this is used, but "all hands brace for
conflicts".

Also note a small change in spelling: we used to write $-foo and
this.myCls('foo'). Now it's this.myCls('foo') and ^foo, which is more
consistent.

The downside to the above is that without an extra (eg. just this.myCls() or ^
in CSS) the class name is foam-u2-md-TextField- with the extra trailing -. Since
we rarely write these by hand, that's not very obvious, but it comes up in the
few places where we reference some other model's CSS name.